Add basic functionalities for digital clock application
A comprehensive foundation for a digital clock application has been added. It comprises the set-up of an Avalonia desktop app with essential modules such as main window view, core files and project settings. Clock time functionalities, for both local and UTC time, have been implemented with an update button. Additionally, a .gitignore file is included to avoid unnecessary files in the repository.